Refactored getExpectedTransactionFromHistory()
I have amended this method based on the time it was taking in a previous test. The test will look through all pages to find the transaction until there are no more pages to find.
This is pointless when the transactions are date ordered. So I am adding a check tha if the transaction is not on the first page there is a good chance an error has occurred.

Additionally, this test includes a minor fix to the recent addition of a checkbox for AD01.