Removed deprecated method
Class.forName("oracle.jdbc.driver.OracleDriver").newInstance(); is deprecated.
In JDBC 4.0 and later, the ServiceLoader mechanism is used to automatically load the JDBC drivers. This means you don't need to explicitly load the JDBC driver using Class.forName(). The JDBC 4.0 drivers include a file named META-INF/services/java.sql.Driver. This file contains the name of the JDBC drivers' implementation classes. When the DriverManager is used to establish a connection, it automatically loads and registers the JDBC drivers.

I've removed the ClassNotFoundException, InstantiationException, and IllegalAccessException from the method signature, as they are no longer needed with this approach.